fix a compile error
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9907 d7cf8633-e32d-0410-b094-e92efae38249

+++ csprogs.h
@@ -57,7 +57,7 @@ extern cvar_t csqc_progsize;
 
 void CL_VM_PreventInformationLeaks(void);
 
-qboolean MakeDownloadPacket(const char *filename, unsigned char *data, unsigned long len, int crc, int cnt, sizebuf_t *buf, int protocol);
+qboolean MakeDownloadPacket(const char *filename, unsigned char *data, size_t len, int crc, int cnt, sizebuf_t *buf, int protocol);
 
 qboolean CL_VM_GetEntitySoundOrigin(int entnum, vec3_t out);
